third_party/WebKit/LayoutTests/fast/css/border-radius-cssText.html

<!DOCTYPE HTML>
<html>
<style>
#target {
   border-radius: 5px;
}
</style>
<div id="target"></div>
<script src="../../resources/js-test.js"></script>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

description("Test for Bug 407012: border-radius shorthand is always splitted to longhand properties in CSSStyleRule.cssText.");

var element = document.getElementById("target");
shouldBe("document.styleSheets[0].rules[0].cssText","'#target { border-radius: 5px; }'");
</script>

</html>
