third_party/WebKit/LayoutTests/fast/css/border-radius-cssText-expected.txt

Test for Bug 407012: border-radius shorthand is always splitted to longhand properties in CSSStyleRule.cssText.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[0].rules[0].cssText is '#target { border-radius: 5px; }'
PASS successfullyParsed is true

TEST COMPLETE
